Move hydrator tests to  GeometriaLabTest/Stdlib/Hydrator
