+++ tests/GeometriaLabTest/Stdlib/Hydrator/FieldsTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace GeometriaLabTest\Stdlib;
+namespace GeometriaLabTest\Stdlib\Hydrator;
 
 use GeometriaLab\Stdlib\Hydrator\Fields;
